+++ src/main/java/com/bp3x/raidbot/RaidBot.java
@@ -2,8 +2,7 @@ package com.bp3x.raidbot;
 
 import com.bp3x.raidbot.commands.lfg.LFGCommand;
 import com.bp3x.raidbot.commands.lfg.LFGReactionListener;
-import com.bp3x.raidbot.commands.LFGCommand;
-import com.bp3x.raidbot.commands.RemindCommand;
+import com.bp3x.raidbot.commands.lfg.RemindCommand;
 import com.bp3x.raidbot.util.Config;
 import com.bp3x.raidbot.util.RaidBotGuildUtils;
 import com.jagrosh.jdautilities.command.CommandClientBuilder;
