+++ src/main/java/com/bp3x/raidbot/commands/lfg/LFGReactionListener.java
@@ -44,7 +44,6 @@ public class LFGReactionListener extends ListenerAdapter {
                     .thenCompose((e) -> reactionEvent.getReaction().removeReaction(reactionEvent.getUser()).submit());
         }
     }
-
     /**
      * Rebuilds the Embed based on the choice processed by handleReactionEvent
      * @param message - the message processed by the bot
